pyproject.toml
[tool.poetry]
name = "whispr"
version = "0.1.0"
description = "social media over signal"
authors = ["technillogue <technillogue@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
forestbot = {git = "https://github.com/mobilecoinofficial/forest", rev = "main"}

[tool.poetry.dev-dependencies]
pylint = "*"
black = "*"
mypy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports=true

disallow_untyped_calls=true
disallow_untyped_defs=true
disallow_untyped_decorators=true # eh
check_untyped_defs=true

warn_redundant_casts=true
warn_unused_ignores=true
warn_unused_configs=true

[tool.pylint.'MESSAGES CONTROL']

disable= [
    "attribute-defined-outside-init",
    "invalid-name",
    "missing-docstring",
    "protected-access",
    "too-few-public-methods",
    "too-many-instance-attributes",
    "logging-fstring-interpolation",
    "consider-using-with",
    "consider-using-from-import",
    "fixme",
    "no-self-use",
    "unspecified-encoding",
    # handled by black
    "format",
    "line-too-long",
]

[tool.pylint.TYPECHECK]
ignored-classes="PrintableWrapper"

[tool.pylint.VARIABLES]
dummy-variables-rgx="(_|dummy|unused|i)"

[tool.pylint.SIMILARITY]
min-similarity-lines=9
ignore-imports="yes"

